Implement role-based access control for posts and users
- Added token verification and user authorization checks for creating, updating, and deleting posts.
- Enhanced user retrieval and management endpoints with access control for admin roles.
- Improved error handling for unauthorized access and missing user data.

+++ src/app/api/posts/[id]/route.ts
@@ -1,3 +1,5 @@
+import { RbacGuard } from "@/app/backend/guard/rbacGuard";
+import { verifyToken } from "@/app/backend/utils/verifyToken";
 import { prisma } from "@/utils/prisma/prisma";
 
 export async function GET(
@@ -23,6 +25,7 @@ export async function GET(
   return Response.json(response);
 }
 
+//Only Admin or Post Author can edit the post
 export async function PUT(
   request: Request,
   { params }: { params: Promise<{ id: string }> }
@@ -35,7 +38,47 @@ export async function PUT(
     return new Response("Error in request body");
   }
 
-  const post = await prisma.post.update({
+  const accessToken = request.headers.get("Authorization")?.split(" ")[1];
+
+  if (!accessToken) {
+    return new Response("Error: No access token");
+  }
+
+  const decoded = verifyToken(accessToken);
+
+  if (typeof decoded === "string") {
+    return new Response("Error: Invalid token");
+  }
+
+  if (!decoded.sub) {
+    return new Response("Error: No user id");
+  }
+
+  const user = await prisma.user.findUnique({
+    where: {
+      id: decoded.sub,
+    },
+  });
+
+  if (!user) {
+    return new Response("Error: User not found");
+  }
+
+  const accessPermission = RbacGuard(user.role, "ADMIN");
+
+  const post = await prisma.post.findUnique({
+    where: { id: id },
+  });
+
+  if (!post) {
+    return new Response("There is no post with the id");
+  }
+
+  if (!accessPermission && user.id !== post.authorId) {
+    return new Response("Error: Unauthorized to edit this post");
+  }
+
+  const updatedPost = await prisma.post.update({
     where: { id: id },
     data: {
       title: body.title,
@@ -43,7 +86,7 @@ export async function PUT(
     },
   });
 
-  if (!post) {
+  if (!updatedPost) {
     return new Response("There is no post with the id");
   }
 
@@ -53,25 +96,67 @@ export async function PUT(
   const response = {
     status: 200,
     statusText: "EditedPostWell",
-    data: post,
+    data: updatedPost,
   };
 
   return Response.json(response);
 }
 
+//Only Admin or Post Author can delete the post
 export async function DELETE(
   request: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const accessToken = request.headers.get("Authorization")?.split(" ")[1];
+
+  if (!accessToken) {
+    return new Response("Error: No access token");
+  }
+
+  const decoded = verifyToken(accessToken);
+
+  if (typeof decoded === "string") {
+    return new Response("Error: Invalid token");
+  }
+
+  const user = await prisma.user.findUnique({
+    where: {
+      id: decoded.sub,
+    },
+  });
+
+  if (!user) {
+    return new Response("Error: User not found");
+  }
+
+  const accessPermission = RbacGuard(user.role, "ADMIN");
+
   const id = (await params).id;
 
-  const index = await prisma.post.delete({
+  const post = await prisma.post.findUnique({
     where: { id: id },
   });
+
+  if (!post) {
+    return new Response("There is no post with the id");
+  }
+
+  if (!accessPermission && user.id !== post.authorId) {
+    return new Response("Error: Unauthorized to delete this post");
+  }
+
+  const deletedPost = await prisma.post.delete({
+    where: { id: id },
+  });
+
+  if (!deletedPost) {
+    return new Response("There is no post with the id");
+  }
 
   const response = {
     status: 200,
     statusText: "DeletedPostWell",
+    data: deletedPost.id,
   };
 
   return Response.json(response);
